feat: restructure authentication components with new styles and layouts for improved usability
