--- src/presentation/widgets/auth/Login/AuthBackground.tsx
@@ -1,5 +1,0 @@
-import styles from "./DesktopLogin.module.css";
-
-export function AuthBackground({ children }: { children: React.ReactNode }) {
-  return <div className={styles.cardBackground}>{children}</div>;
-}
