--- src/presentation/widgets/auth/Login/DesktopLogin.module.css
@@ -1,43 +1,0 @@
-.cardBackground {
-  position: relative;
-  width: 680px;
-  min-height: 400px;
-  border-radius: 12px;
-  background-color: var(--bg-card);
-  padding: 170px 40px 48px;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-}
-
-.search {
-  position: absolute;
-  top: 48px;
-  left: 57px;
-}
-
-.task {
-  position: absolute;
-  top: 91.96px;
-  left: 577.6px;
-}
-
-.title {
-  max-width: 300px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  text-align: center;
-  white-space: pre-line;
-  gap: 12px;
-  position: absolute;
-  top: 40px;
-  left: 50%;
-  transform: translateX(-50%);
-}
-
-.formWapper {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-}
