--- src/presentation/widgets/auth/Login/index.ts
@@ -1,1 +1,0 @@
-export { DesktopLogin } from "./DesktopLogin";
