+++ src/presentation/widgets/auth/LoginForm/index.ts
@@ -1,0 +1,2 @@
+export { LoginFormDesktop } from "./LoginFormDesktop";
+export { LoginFormMobile } from "./LoginFormMobile";
