+++ src/presentation/widgets/auth/_components/AuthBackground.module.css
@@ -1,0 +1,11 @@
+.background {
+  position: relative;
+  width: 680px;
+  min-height: 400px;
+  border-radius: 12px;
+  background-color: var(--bg-card);
+  padding: 170px 40px 48px;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+}
