+++ src/presentation/widgets/auth/_components/AuthBackground.tsx
@@ -1,0 +1,5 @@
+import styles from "./AuthBackground.module.css";
+
+export function AuthBackground({ children }: { children: React.ReactNode }) {
+  return <div className={styles.background}>{children}</div>;
+}
