+++ src/presentation/widgets/auth/_components/BottomContent.module.css
@@ -1,0 +1,9 @@
+.bottomContent {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+  position: absolute;
+  bottom: 32px;
+  left: 50%;
+  transform: translateX(-50%);
+}
