+++ src/presentation/widgets/auth/_components/BottomContent.tsx
@@ -1,0 +1,5 @@
+import styles from "./BottomContent.module.css";
+
+export function BottomContent({ children }: { children: React.ReactNode }) {
+  return <div className={styles.bottomContent}>{children}</div>;
+}
