+++ src/presentation/widgets/auth/_components/TitleWrapper.module.css
@@ -1,0 +1,12 @@
+.title {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  white-space: pre-line;
+  gap: 12px;
+  position: absolute;
+  top: 40px;
+  left: 50%;
+  transform: translateX(-50%);
+}
