+++ src/presentation/widgets/auth/_components/TitleWrapper.tsx
@@ -1,0 +1,5 @@
+import styles from "./TitleWrapper.module.css";
+
+export function TitleWrapper({ children }: { children: React.ReactNode }) {
+  return <div className={styles.title}>{children}</div>;
+}
