+++ src/presentation/widgets/auth/_components/index.ts
@@ -1,0 +1,5 @@
+export { AuthBackground } from "./AuthBackground";
+export { TitleWrapper } from "./TitleWrapper";
+export { BottomContent } from "./BottomContent";
+export { Input } from "./Input";
+export { SubmitButton } from "./SubmitButton";
